rename qna component folders
