+++ src/components/qna-components/answer/qna-answer.component.jsx
