+++ src/components/qna-components/list/qna-list.styles.scss
