+++ src/components/qna-components/question/qna-question.styles.scss
